pyproject.toml
[project]
name = "droneserver"
version = "0.2.0"
description = "MAVLink MCP Server - AI-powered drone control via Model Context Protocol"
authors = [
    {name = "Peter J Burke"},
    {name = "Ion Gabriel", email = "ion.ion.gabriel@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "mavsdk>=2.0.0",
    "mcp-agent>=0.1.0",
    "mcp-server>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/PeterJBurke/droneserver"
Repository = "https://github.com/PeterJBurke/droneserver"
"Original Repository" = "https://github.com/ion-g-ion/MAVLinkMCP"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src"]
